imbue_common: consolidate placeholder ANTHROPIC_API_KEY into a single constant
Both apps/remote_service_connector/scripts/create_pool_hosts.py and
apps/minds/imbue/minds/desktop_client/agent_creator.py defined their own
PLACEHOLDER_ANTHROPIC_API_KEY constant with the same value. The values
happened to match but nothing prevented drift, and only the connector
side actually used it -- the minds-side definition was only referenced
by two characterization tests in agent_creator_test.py.

Move the single source of truth to libs/imbue_common/imbue/imbue_common/
pool_host_constants.py, where both apps already meet (they both depend
on imbue-common as a workspace dep). The connector script imports from
there; the minds desktop client no longer carries a redundant copy. The
characterization tests move alongside the constant in
imbue_common/pool_host_constants_test.py.
