+++ apps/minds/imbue/minds/desktop_client/agent_creator.py
@@ -93,16 +93,6 @@ OutputCallback = Callable[[str, bool], None]
 
 LOG_SENTINEL: Final[str] = "__DONE__"
 
-# Placeholder ANTHROPIC_API_KEY set on pre-created pool hosts. Must look like
-# a real Anthropic key (correct prefix, correct length ~108 chars) so that
-# Claude Code validation accepts it during provisioning. The real LiteLLM
-# virtual key is written into the agent env file and claude config during
-# lease setup (see _build_inject_anthropic_command and
-# _build_patch_claude_config_command).
-PLACEHOLDER_ANTHROPIC_API_KEY: Final[str] = (
-    "sk-ant-api03-PLACEHOLDER000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-)
-
 
 @pure
 def _build_inject_anthropic_command(
